// src/app/core/models/vehicle-ads.ts
export interface VehicleAd {
    images: string[];
    videos: string[];
    title: string;
    category: string;
    subCategory: string;
    location: string;
    price: number;
    modelYear: number;
    registeredIn: string;
    make: string;
    engineType: string;
    description: string;
    localOrImported: string;
    postedBy: string;
    adType: string;
    chatCount: number;
    callCount: number;
    favorites: string[];
    sold: boolean;
    _id: string;
    createdAt: string;
    views: string[];
    __v: number;
  }
  